+++ docs/changelog.rst
@@ -37,6 +37,11 @@ New features
   2. Semicolon followed by a space
   3. Comma followed by a space
   4. Slash wrapped by spaces
+- :doc:`plugins/smartplaylist`: Add new configuration option ``dest_regen`` to
+  regenerate items' path in the generated playlist instead of using the ones of
+  the library. This is useful when items have been imported in don't copy-move
+  (`-C -M`) mode in the library but are later passed through the `convert`
+  plugin which will regenerate new paths according to the Beets path format.
 
 Bug fixes
 ~~~~~~~~~